Reverted on the flexible search for now..    "Decentralized finance news"
The time duration still is allowed.

Also I used pprint, PrettyPrint to make the stdout easy to read
but still log it CSV at the moment.

I need to think through this more..

The was to assit me and writing up reports on a given currency.

My goal was to collect data on a given Crypto currency/token:
   1. Price information - other script basically - collect coingecko over time.
        * look for trends and changes and compare with other currencies.
        * Look for news to correlate with price changes.

   2. Recent and trends in news on a Crypto coin - actually summary and link.
       - What is currently happening in the news for this coin/token.
       - What are the long term trends on news items for this.
       - Long term statistics on the frequency of news.